Update 1.package_and_multisign.py
Nova atualização 06/11/2024
Agora sinconiza com todos os javas Google.

+++ 1.package_and_multisign.py
@@ -3,9 +3,21 @@ import re
 import os
 import time
 import subprocess
+import requests
 import shutil
 import glob
 import sys 
+
+
+variavelcomalias = "canalqb"
+variavelcomnomedokeystore = f"{variavelcomalias}.keystore"
+variavelsenha = "xxxx" #Senha com 6 caracteres
+
+if len(variavelsenha) != 6:
+    print('Coloque uma senha de 6 caracteres na variavel variavelsenha que está na linha 14')
+    print('Enter a 6-character password in the variable variavelsenha on line 14')
+
+    exit()  # Sai do script
 
 try:
     command = "taskkill /IM java.exe /F"
@@ -54,11 +66,8 @@ def rename_file(option):
     # Renomeia o arquivo
     os.rename(original_file_path, new_file_name)
     print(f"Arquivo renomeado para: {os.path.basename(new_file_name)}")  
-    return new_file_name
+    return new_file_name 
     
-
-
-
 # Função para esperar até o arquivo ser criado
 def wait_for_file(file_path):
     while not os.path.exists(file_path):
@@ -327,9 +336,10 @@ print("Empacotamento concluído.")
 
 #nome_final_do_arquivo = os.path.basename(rename_file(opcao))
 nome_final_do_arquivo = rename_file(opcao)
-
+os.system('cls')
 # Caminho de origem (ajuste conforme necessário) 
 origem_simbolos = os.path.join('build', app_build, 'android', 'gradle', 'app', 'build', 'intermediates', 'merged_native_libs', 'release', 'out', 'lib')
+os.system('cls')
 if opcao == '1':
     nomepasta = '-mt.st.0.0.1.aab'
 elif opcao == '2':
@@ -450,3 +460,143 @@ if os.path.exists(diretorio_para_compactar):
         print('Erro ao criar o arquivo .zip.')
 else:
     print(f'Erro: O diretório {diretorio_para_compactar} não existe.')
+     
+# Variáveis do arquivo e keystore
+variavelcomnomedoarquivo = os.path.basename(nome_final_do_arquivo)
+
+
+# Diretório onde os arquivos serão salvos
+DIST_DIR = 'dist'
+
+# Função para rodar um comando no sistema
+def run_command(command, cwd=None):
+    try:
+        result = subprocess.run(command, shell=True, check=True, text=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=cwd)
+        print(result.stdout)
+        print(result.stderr)
+    except subprocess.CalledProcessError as e:
+        print(f"Erro ao executar o comando: {e}")
+        print(e.output)
+        print(e.stderr)
+
+# Função para obter o nome do usuário de forma compatível com diferentes sistemas operacionais
+def get_current_user():
+    return os.getenv('USERNAME') if os.name == 'nt' else os.getenv('USER') or os.getenv('LOGNAME')
+
+# Função para verificar e adicionar um diretório ao PATH se necessário
+def verify_and_add_to_path(directory, description):
+    if os.path.isdir(directory):
+        print(f'{description} encontrado em {directory}')
+        if directory not in os.environ['PATH']:
+            if input(f'O diretório {directory} não está no PATH. Deseja adicionar? (s/n): ').lower() == 's':
+                os.environ['PATH'] += f";{directory}"
+                print(f'{directory} adicionado ao PATH.')
+            else:
+                print(f'{directory} não foi adicionado ao PATH.')
+    else:
+        print(f'{description} não encontrado. Por favor, atualize o BeeWare.')
+
+# Função para verificar as versões e dependências
+def check_dependencies(env_user):
+    java_path = f'C:\\Users\\{env_user}\\AppData\\Local\\BeeWare\\briefcase\\Cache\\tools\\java17\\bin'
+    android_sdk_path = f'C:\\Users\\{env_user}\\AppData\\Local\\BeeWare\\briefcase\\Cache\\tools\\android_sdk\\build-tools\\34.0.0'
+    verify_and_add_to_path(java_path, 'Java 17')
+    verify_and_add_to_path(android_sdk_path, 'Android SDK build-tools')
+
+# Função para baixar o BundleTool
+def download_bundletool():
+    bundletool_url = 'https://github.com/google/bundletool/releases/download/1.16.0/bundletool-all-1.16.0.jar'
+    
+    # Garantir que o diretório dist existe
+    os.makedirs(DIST_DIR, exist_ok=True)
+    
+    dist_path = os.path.join(DIST_DIR, 'bundletool-all-1.16.0.jar')
+    print(f'Downloading BundleTool para {dist_path}')
+    
+    response = requests.get(bundletool_url)
+    with open(dist_path, 'wb') as file:
+        file.write(response.content)
+    
+    print('Download do BundleTool concluído.')
+
+# Função para realizar a assinatura do APK
+def sign_apk(arquivo, keystore, alias, senha_keystore): 
+    command = f'apksigner sign --ks "{keystore}" --ks-pass pass:{senha_keystore} --min-sdk-version 34 "{arquivo}"'
+    run_command(command, cwd=DIST_DIR)
+
+# Função para validar o arquivo bundle
+def validate_bundle(arquivo):
+    command = f'java -jar bundletool-all-1.16.0.jar validate --bundle "{arquivo}"' 
+    run_command(command, cwd=DIST_DIR)
+
+# Função para gerar o arquivo zip com a chave de criptografia
+def generate_encryption_zip(keystore, alias, senha_keystore):
+    print("Baixando o pepk.jar...")
+
+    # URL do pepk.jar
+    pepk_url = "https://github.com/canalqb/briefcase/raw/refs/heads/main/pepk.jar"
+    
+    # Baixando pepk.jar diretamente para o diretório dist
+    pepk_path = os.path.join(DIST_DIR, 'pepk.jar')
+    response = requests.get(pepk_url)
+    with open(pepk_path, 'wb') as file:
+        file.write(response.content)
+
+    # Garantir que o arquivo pepk.jar foi baixado corretamente
+    pepk_path_abs = os.path.abspath(pepk_path)
+    if not os.path.exists(pepk_path_abs):
+        print(f"Erro: O arquivo pepk.jar não foi encontrado em {pepk_path_abs}")
+        return
+    
+    while not os.path.exists(os.path.join(DIST_DIR, 'encryption_public_key.pem')):
+        print("Acesse o Google Play Console para baixar sua pem atual e coloque na pasta dist.")
+        time.sleep(2)
+    
+    print("Executando o pepk.jar para gerar o arquivo output.zip")
+
+    # Caminho absoluto para o arquivo de chave pública
+    encryption_key_path = os.path.abspath(os.path.join(DIST_DIR, 'encryption_public_key.pem'))
+
+    # Comando para rodar o pepk.jar com caminhos absolutos
+    command = f'java -jar "{pepk_path_abs}" --keystore="{keystore}" --alias="{alias}" --output=output.zip --include-cert --rsa-aes-encryption --encryption-key-path="{encryption_key_path}" --keystore-pass="{senha_keystore}"'
+    
+    run_command(command, cwd=DIST_DIR)
+    print("Arquivo output.zip gerado com sucesso.")
+
+# Função para garantir que o diretório existe
+def ensure_dist_dir_exists():
+    abs_dist_dir = os.path.abspath(DIST_DIR)
+    if not os.path.exists(abs_dist_dir):
+        os.makedirs(abs_dist_dir)
+        print(f'Diretório {abs_dist_dir} criado com sucesso.')
+
+# Identificar o usuário atual no Windows
+env_user = get_current_user()
+print(f'O usuário atual é: {env_user}')
+
+# Verificar as dependências
+check_dependencies(env_user)
+
+# Caminho completo do keystore (usando caminho absoluto)
+ensure_dist_dir_exists()  # Verifica e cria o diretório 'dist'
+keystore_path = os.path.join(os.path.abspath(DIST_DIR), variavelcomnomedokeystore).replace("\\", "/")
+
+print(f"Gerando chave do keystore em {keystore_path}")
+# Comando para gerar a chave
+command = f'keytool -genkeypair -v -keystore "{keystore_path}" -alias canalqb -keyalg RSA -keysize 2048 -validity 10000 -storepass 278307 -dname "CN=canalqb, OU=canalqb, O=canalqb, L=Sao Paulo, ST=Sao Paulo, C=BR"'
+run_command(command, cwd=os.path.abspath(DIST_DIR))
+
+# Baixar o BundleTool
+download_bundletool()
+
+# Validar o arquivo .aab com o BundleTool
+validate_bundle(variavelcomnomedoarquivo)
+
+# Assinar o APK
+sign_apk(variavelcomnomedoarquivo, keystore_path, variavelcomalias, variavelsenha)
+
+# Validar novamente
+validate_bundle(variavelcomnomedoarquivo)
+
+# Gerar o arquivo zip com a chave de criptografia
+generate_encryption_zip(keystore_path, variavelcomalias, variavelsenha)
